requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
psycopg2-binary
python-jose
python-multipart
bcrypt
passlib[bcrypt]
alembic
pydantic[email]
python-dotenv
email-validator
stripe
